fix issues, removed redundant folders
