+++ components/bl-snackbar-component/src/button.js
@@ -1,0 +1,5 @@
+export function Action({ className, onClick, children }) {
+  return (
+    <button className={ className } onClick={ onClick }> { children } </button>
+  )
+}
