--- components/bl-snackbar-component/src/components/button.js
@@ -1,9 +1,0 @@
-export const Action = ({ className, onClick, children }) => {
-  return (
-    <button
-      className={ className }
-      onClick={ onClick }>
-      { children }
-    </button>
-  )
-}
