--- components/bl-snackbar-component/src/components/icons/index.js
@@ -1,5 +1,0 @@
-export * from './error'
-export * from './info'
-export * from './success'
-export * from './warning'
-export * from './close'
