--- components/bl-snackbar-component/src/helpers/type.js
@@ -1,8 +1,0 @@
-import { Error, Info, Success, Warning } from '../components/icons'
-
-export const IconsMap = {
-  error  : Error,
-  success: Success,
-  warning: Warning,
-  info   : Info,
-}
